scripts/helm-check-for-updates/requirements.txt
pyyaml
subprocess